feat: Add raw API functions for config-manager. Add a few config-manager related flags to existing functions. (#522)

+++ src/utils/ExportImportUtils.ts
@@ -42,12 +42,14 @@ export type ExportImport = {
    * @param {Object} data data object
    * @param {String} filename file name
    * @param {boolean} includeMeta true to include metadata, false otherwise. Default: true
+   * @param {boolean} keepRev keep the _rev key from objects. Default: false
    * @return {boolean} true if successful, false otherwise
    */
   saveJsonToFile(
     data: object,
     filename: string,
-    includeMeta?: boolean
+    includeMeta?: boolean,
+    keepRev?: boolean
   ): boolean;
   /**
    * Save text data to file
@@ -176,9 +178,10 @@ export default (state: State): ExportImport => {
     saveJsonToFile(
       data: object,
       filename: string,
-      includeMeta = true
+      includeMeta = true,
+      keepRev = false
     ): boolean {
-      return saveJsonToFile({ data, filename, includeMeta, state });
+      return saveJsonToFile({ data, filename, includeMeta, keepRev, state });
     },
     saveTextToFile(data: string, filename: string): boolean {
       return saveTextToFile({ data, filename, state });
@@ -387,24 +390,29 @@ export function saveToFile({
  * @param {object} data data object
  * @param {string} filename file name
  * @param {boolean} includeMeta true to include metadata, false otherwise. Default: true
+ * @param {boolean} keepRev Keep the _rev key from objects. Default: false
  * @return {boolean} true if successful, false otherwise
  */
 export function saveJsonToFile({
   data,
   filename,
   includeMeta = true,
+  keepRev = false,
   state,
 }: {
   data: object;
   filename: string;
   includeMeta?: boolean;
+  keepRev?: boolean;
   state: State;
 }): boolean {
   const exportData = data;
   if (includeMeta && !exportData['meta'])
     exportData['meta'] = getMetadata({ state });
   if (!includeMeta && exportData['meta']) delete exportData['meta'];
-  deleteDeepByKey(exportData, '_rev');
+  if (!keepRev) {
+    deleteDeepByKey(exportData, '_rev');
+  }
   return saveTextToFile({
     data: stringify(exportData),
     filename,
